+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add Sentry.logger examples to example pages when logs are enabled ([#1127](https://github.com/getsentry/sentry-wizard/pull/1127))
+
 ## 6.8.0
 
 ### Features